refactor: Make chart version updater functionality public
