+++ hack/release/cmd/prerelease/prerelease.go
@@ -4,22 +4,16 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"path/filepath"
-	"strings"
 
-	"github.com/drone/envsubst"
+	"github.com/mesosphere/kommander-applications/hack/release/pkg/chartversion"
 	"github.com/spf13/cobra"
 )
 
 var Cmd *cobra.Command //nolint:gochecknoglobals // Cobra commands are global.
 
 const (
-	versionFlagName               = "version"
-	repoFlagName                  = "repo"
-	kommanderChartVersionTemplate = "${kommanderChartVersion:=%s}"
-
-	kommanderHelmReleasePathPattern        = "./services/kommander/*/kommander.yaml"
-	kommanderAppMgmtHelmReleasePathPattern = "./services/kommander-appmanagement/*/kommander-appmanagement.yaml"
+	versionFlagName = "version"
+	repoFlagName    = "repo"
 )
 
 func init() { //nolint:gochecknoinits // Initializing cobra application.
@@ -33,8 +27,8 @@ func init() { //nolint:gochecknoinits // Initializing cobra application.
 			if _, err := os.Stat(kommanderApplicationsRepo); os.IsNotExist(err) {
 				return err
 			}
-			fullChartVersion := fmt.Sprintf(kommanderChartVersionTemplate, chartVersion)
-			err := updateChartVersions(kommanderApplicationsRepo, fullChartVersion)
+
+			err := chartversion.UpdateChartVersions(kommanderApplicationsRepo, chartVersion)
 			if err != nil {
 				return err
 			}
@@ -54,48 +48,4 @@ func init() { //nolint:gochecknoinits // Initializing cobra application.
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-func updateChartVersions(kommanderApplicationsRepo, chartVersion string) error {
-	kommanderHelmReleasePaths := []string{kommanderHelmReleasePathPattern, kommanderAppMgmtHelmReleasePathPattern}
-	for _, helmReleasePath := range kommanderHelmReleasePaths {
-		// Find the HelmRelease
-		matches, err := filepath.Glob(filepath.Join(kommanderApplicationsRepo, helmReleasePath))
-		if err != nil {
-			return err
-		}
-		if len(matches) == 0 {
-			return fmt.Errorf("no matches found for HelmRelease path %s (verify the kommander-applications repo path is correct)", helmReleasePath)
-		}
-		if len(matches) > 1 {
-			return fmt.Errorf("found > 1 match for HelmRelease path %s (there should only be one match)", helmReleasePath)
-		}
-		helmReleaseFilePath := matches[0]
-
-		// Update the kommanderChartVersion value
-		parsedFile, err := envsubst.ParseFile(helmReleaseFilePath)
-		if err != nil {
-			return err
-		}
-		subVars := map[string]string{
-			"kommanderChartVersion": chartVersion,
-			"releaseNamespace":      "${releaseNamespace}",
-		}
-		updatedFile, err := parsedFile.Execute(func(s string) string {
-			return subVars[s]
-		})
-		if err != nil {
-			return err
-		}
-
-		if !strings.Contains(updatedFile, chartVersion) {
-			return fmt.Errorf("failed to update Kommander HelmRelease chart version")
-		}
-
-		err = os.WriteFile(helmReleaseFilePath, []byte(updatedFile), 0644)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }
